requirements.txt
boto
PyGitHub
psycopg2-binary
sqlalchemy 
python-dotenv
